remove withDockerCompose
