+++ dev-packages/node-integration-tests/utils/runner/createRunner.ts
@@ -14,8 +14,7 @@ import type {
 } from '@sentry/core';
 import { normalize } from '@sentry/core';
 import { createBasicSentryServer } from '@sentry-internal/test-utils';
-import { spawn, spawnSync } from 'child_process';
-import { createHash } from 'crypto';
+import { spawn } from 'child_process';
 import { existsSync } from 'fs';
 import { tmpdir } from 'os';
 import { join } from 'path';
@@ -34,13 +33,6 @@ import {
   assertSentryTransaction,
   assertSpanEnvelopeHeader,
 } from './../assertions';
-
-export interface DockerOptions {
-  /**
-   * The working directory to run docker compose in
-   */
-  workingDirectory: string[];
-}
 
 type VoidFunction = () => void;
 
@@ -142,7 +134,6 @@ export function createRunner(...paths: string[]) {
   let unordered = false;
   let withEnv: Record<string, string> = {};
   let withSentryServer = false;
-  let dockerOptions: DockerOptions | undefined;
   let ensureNoErrorOutput = false;
   // When set, the test using this runner expects `completed()` to reject (e.g. `test.fails` variants
   // created via `createEsmAndCjsTests` with `failsOnEsm`/`failsOnCjs`). We suppress the captured-log
@@ -229,10 +220,6 @@ export function createRunner(...paths: string[]) {
     },
     unordered: function () {
       unordered = true;
-      return this;
-    },
-    withDockerCompose: function (options: DockerOptions) {
-      dockerOptions = options;
       return this;
     },
     ensureNoErrorOutput: function () {
@@ -397,28 +384,17 @@ export function createRunner(...paths: string[]) {
       // We need to properly define & pass these types around for TS 3.8,
       // which otherwise fails to infer these correctly :(
       type ServerStartup = [number | undefined, (() => void) | undefined];
-      type DockerStartup = VoidFunction | undefined;
 
       const serverStartup: Promise<ServerStartup> = withSentryServer
         ? createBasicSentryServer(newEnvelope)
         : Promise.resolve([undefined, undefined]);
 
-      const dockerStartup: Promise<DockerStartup> = dockerOptions
-        ? runDockerCompose(dockerOptions)
-        : Promise.resolve(undefined);
-
-      const startup = Promise.all([dockerStartup, serverStartup]);
-
-      startup
-        .then(([dockerChild, [mockServerPort, mockServerClose]]) => {
+      serverStartup
+        .then(([mockServerPort, mockServerClose]) => {
           if (mockServerClose) {
             registerCleanupStep(() => {
               mockServerClose();
             });
-          }
-
-          if (dockerChild) {
-            registerCleanupStep(dockerChild);
           }
 
           const env = mockServerPort
@@ -633,68 +609,6 @@ export function createRunner(...paths: string[]) {
       };
     },
   };
-}
-
-/**
- * Runs `docker compose up -d --wait`, which blocks until every service's
- * healthcheck reports healthy. Each suite defines its healthcheck in its
- * own docker-compose.yml.
- *
- * Returns a function that can be called to docker compose down
- */
-export async function runDockerCompose(options: DockerOptions): Promise<VoidFunction> {
-  const cwd = join(...options.workingDirectory);
-
-  // Docker Compose derives the project name from the compose file's directory
-  // basename by default. Several suites live in directories that share a
-  // basename (e.g. `tracing/mysql2` and `tracing/knex/mysql2`), so they collide
-  // on the same project + network when running in parallel: one suite's
-  // teardown removes the shared `<name>_default` network while a sibling is
-  // still starting, producing "network <name>_default not found". Deriving a
-  // unique, stable project name from the full working directory isolates every
-  // suite from each other.
-  const projectName = `sentry-it-${createHash('sha1').update(cwd).digest('hex').slice(0, 12)}`;
-  const composeArgs = (...args: string[]): string[] => ['compose', '-p', projectName, ...args];
-
-  const close = (): void => {
-    spawnSync('docker', composeArgs('down', '--volumes'), {
-      cwd,
-      stdio: process.env.DEBUG ? 'inherit' : undefined,
-    });
-  };
-
-  // ensure we're starting fresh
-  close();
-
-  const composeUp = (): ReturnType<typeof spawnSync> =>
-    spawnSync('docker', composeArgs('up', '-d', '--wait'), {
-      cwd,
-      stdio: process.env.DEBUG ? 'inherit' : 'pipe',
-    });
-
-  // `docker compose up` occasionally fails on CI with transient daemon races
-  // (e.g. "failed to set up container networking: network <x>_default not
-  // found" right after the network was created). A clean teardown plus retry
-  // clears these, while genuine healthcheck failures stay red on every attempt.
-  const maxAttempts = 3;
-  let result = composeUp();
-  for (let attempt = 1; attempt < maxAttempts && result.status !== 0; attempt++) {
-    close();
-    result = composeUp();
-  }
-
-  if (result.status !== 0) {
-    const stderr = result.stderr?.toString() ?? '';
-    const stdout = result.stdout?.toString() ?? '';
-    // Surface container logs to make healthcheck failures easier to diagnose in CI
-    const logs = spawnSync('docker', composeArgs('logs'), { cwd }).stdout?.toString() ?? '';
-    close();
-    throw new Error(
-      `docker compose up --wait failed (exit ${result.status})\n${stderr}${stdout}\n--- container logs ---\n${logs}`,
-    );
-  }
-
-  return close;
 }
 
 interface Deferred {
